Identicon (identicon v0.1.0)

Documentation for Identicon.

Functions

build_grid(image_struct)

Builds a grid of colors to display as Identicon

Parameters: the hexadecimal representation of grid colors from our image_struct

build_pixel_map(image_struct)

draw_image(image)

filter_out_odd_squares(image_struct)

hash_input(input_str)

Returns an image struct representation of the hashed value of input_str

main(input_str)

Main function to create our identicon from an input string

Examples

iex> Identicon.main("Asdf")
:ok

Now, find a file named "Asdf.png" at the root of your project

mirror_row(row)

pick_color(image_struct)

save_image(image_file, input_str)