identicon v0.1.0 Identicon

Generates an identicon based on a string

Functions

build_grid(image)

Returns an Identicon.Image struct, containing a list of integers lists representing a grid. The image is a Identicon.Image struct to generate a grid.

Example

iex> image = Identicon.hash_string("identicon")
iex> Identicon.build_grid(image)
%Identicon.Image{
  color: {},
  grid: [
    {43, 0}, {65, 1}, {97, 2}, {65, 3}, {43, 4},
    {60, 5}, {135, 6}, {2, 7}, {135, 8}, {60, 9},
    {181, 10}, {55, 11}, {43, 12}, {55, 13}, {181, 14},
    {189, 15}, {201, 16}, {168, 17}, {201, 18}, {189, 19},
    {16, 20}, {112, 21}, {64, 22}, {112, 23}, {16, 24}
  ],
  hex: [173, 43, 65, 97, 60, 135, 2, 181, 55, 43, 189, 201, 168, 16, 112, 64],
  painted_cells: [],
  pixel_map: {}
}

build_pixel_map(image)

Returns an Identicon.Image struct, containing a pixel map representing squares to be colored. The image is a Identicon.Image struct to transform painted_cells in pixel_map.

Example

iex> image = Identicon.hash_string("identicon")
iex> image = Identicon.build_grid(image)
iex> image = Identicon.filter_odd_squares(image)
iex> Identicon.build_pixel_map(image)
%Identicon.Image{
  color: {},
  grid: [
    {43, 0}, {65, 1}, {97, 2}, {65, 3}, {43, 4},
    {60, 5}, {135, 6}, {2, 7}, {135, 8}, {60, 9},
    {181, 10}, {55, 11}, {43, 12}, {55, 13}, {181, 14},
    {189, 15}, {201, 16}, {168, 17}, {201, 18}, {189, 19},
    {16, 20}, {112, 21}, {64, 22}, {112, 23}, {16, 24}
  ],
  hex: [173, 43, 65, 97, 60, 135, 2, 181, 55, 43, 189, 201, 168, 16, 112, 64],
  painted_cells: [
    {60, 5}, {2, 7}, {60, 9},
    {168, 17}, {16, 20}, {112, 21},
    {64, 22}, {112, 23}, {16, 24}
  ],
  pixel_map: [
    {{2, 52}, {48, 98}},
    {{102, 52}, {148, 98}},
    {{202, 52}, {248, 98}},
    {{102, 152}, {148, 198}},
    {{2, 202}, {48, 248}},
    {{52, 202}, {98, 248}},
    {{102, 202}, {148, 248}},
    {{152, 202}, {198, 248}},
    {{202, 202}, {248, 248}}
  ]
}

draw_image(image)

Returns a raw image (generated by Erlang’s :egd). The image is a Identicon.Image struct to generate a raw image

filter_odd_squares(image)

Returns an Identicon.Image struct, containing a filtered (by even) list of integers lists representing a grid. The image is a Identicon.Image struct to filter even integers from a grid data.

Example

iex> image = Identicon.hash_string("identicon")
iex> image = Identicon.build_grid(image)
iex> Identicon.filter_odd_squares(image)
%Identicon.Image{
  color: {},
  grid: [
    {43, 0}, {65, 1}, {97, 2}, {65, 3}, {43, 4},
    {60, 5}, {135, 6}, {2, 7}, {135, 8}, {60, 9},
    {181, 10}, {55, 11}, {43, 12}, {55, 13}, {181, 14},
    {189, 15}, {201, 16}, {168, 17}, {201, 18}, {189, 19},
    {16, 20}, {112, 21}, {64, 22}, {112, 23}, {16, 24}
  ],
  hex: [173, 43, 65, 97, 60, 135, 2, 181, 55, 43, 189, 201, 168, 16, 112, 64],
  painted_cells: [
    {60, 5},
    {2, 7},
    {60, 9},
    {168, 17},
    {16, 20},
    {112, 21},
    {64, 22},
    {112, 23},
    {16, 24}
  ],
  pixel_map: {}
}

get_coordinates(size, index)

Returns a tuple of tuples, representing two coordinates (x, y) in space. The size means “how many cells has in this space/plane”. The index means the cell position in this space.

Example

Get coordinate of X in this plane: .... |||| |||X| |||_|

iex> Identicon.get_coordinates(3, 5)
{{102, 52}, {148, 98}}

hash_string(input)

Returns an Identicon.Image struct, containing hex as a list of integers representing hash. The input is a string to calculate hash.

Example

iex> Identicon.hash_string("identicon")
%Identicon.Image{
  color: {},
  hex: [173, 43, 65, 97, 60, 135, 2, 181, 55, 43, 189, 201, 168, 16, 112, 64]
}

main(input)

mirror_row(row)

Returns a row mirrored by the last element. The row is any list of data

Example

iex> Identicon.mirror_row([1, 2, 3])
[1, 2, 3, 2, 1]

pick_color(image)

Returns an Identicon.Image struct, containing a RGB color as a list of integers. The image is a Identicon.Image struct to calculate RGB color.

Example

iex> image = Identicon.hash_string("identicon")
iex> Identicon.pick_color(image)
%Identicon.Image{
  color: {173, 43, 65},
  grid: [],
  hex: [173, 43, 65, 97, 60, 135, 2, 181, 55, 43, 189, 201, 168, 16, 112, 64],
  painted_cells: [],
  pixel_map: {}
}

save_image(image, path)

Returns a tuple {status, result}. Status could be both: :ok or :error. Result could be both: path or Something went wrong The image is a raw image (generated by Erlang’s :egd). The path is a absolute path to where image will be saved.