identicon v0.1.0 Identicon

Generates an identicon based on a string

Functions

build_grid(image)

Returns an Identicon.Image struct, containing a list of integers lists representing a grid,

Example

iex> image = Identicon.hash_string("identicon")
iex> Identicon.build_grid(image)
%Identicon.Image{
  color: %Identicon.Color{blue: 0, green: 0, red: 0},
  grid: [
    {43, 0}, {65, 1}, {97, 2}, {65, 3}, {43, 4},
    {60, 5}, {135, 6}, {2, 7}, {135, 8}, {60, 9},
    {181, 10}, {55, 11}, {43, 12}, {55, 13}, {181, 14},
    {189, 15}, {201, 16}, {168, 17}, {201, 18}, {189, 19},
    {16, 20}, {112, 21}, {64, 22}, {112, 23}, {16, 24}
  ],
  hex: [173, 43, 65, 97, 60, 135, 2, 181, 55, 43, 189, 201, 168, 16, 112, 64]
}

filter_odd_squares(image)

Returns an Identicon.Image struct, containing a filtered (by even) list of integers lists representing a grid.

Example

iex> image = Identicon.hash_string("identicon")
iex> grid = Identicon.build_grid(image)
iex> Identicon.filter_odd_squares(grid)
%Identicon.Image{
  color: %Identicon.Color{blue: 0, green: 0, red: 0},
  grid: [
    {60, 5},
    {2, 7},
    {60, 9},
    {168, 17},
    {16, 20},
    {112, 21},
    {64, 22},
    {112, 23},
    {16, 24}
  ],
  hex: [173, 43, 65, 97, 60, 135, 2, 181, 55, 43, 189, 201, 168, 16, 112, 64]
}

hash_string(input)

Returns an Identicon.Image struct, containing hex as a list of integers representing hash. The input is a string to calculate hash.

Example

iex> Identicon.hash_string("identicon")
%Identicon.Image{
  color: %Identicon.Color{blue: 0, green: 0, red: 0},
  hex: [173, 43, 65, 97, 60, 135, 2, 181, 55, 43, 189, 201, 168, 16, 112, 64]
}

main(input)

mirror_row(row)

Returns a row mirrored by the last element,

Example

iex> Identicon.mirror_row([1, 2, 3])
[1, 2, 3, 2, 1]

pick_color(image)

Returns an Identicon.Image struct, containing a RGB color as a list of integers. The image is a Identicon.Image struct to calculate RGB color.

Example

iex> image = Identicon.hash_string("identicon")
iex> Identicon.pick_color(image)
%Identicon.Image{
  color: %Identicon.Color{blue: 65, green: 43, red: 173},
  hex: [173, 43, 65, 97, 60, 135, 2, 181, 55, 43, 189, 201, 168, 16, 112, 64]
}