Identicon (identicon v0.1.0)

Creates an identicon for a given text string

Functions

build_pixel_map(image)

create_grid(image)

draw_image(image)

filter_odds(grid)

Specs

filter_odds({integer(), any()}) :: [{integer(), any()}]

Basic filter to check if code integer is divisible by two (even) inside the tuple grid structure {code, index}

Examples

iex> filtered_list = Identicon.filter_odds([{1, 1}, {245, 2}, {244, 15}])
[{244, 15}]
iex> filtered_list
[{244, 15}]
iex> length filtered_list
1

hash_input(input)

Uses the crypto and binary Erlang libraries to hash text string

Examples

iex> Identicon.hash_input("asdf")
%Identicon.Image{
  color: nil,
  grid: nil,
  hex: [145, 46, 200, 3, 178, 206, 73, 228, 165, 65, 6, 141, 73, 90, 181, 112],
  pixel_map: nil
}

pick_color(image)

reflect_columns(grid_row)

We assume a 5x5 grid. Each row is a list of three numbers (the first three columns). The first and second need to be reflected a, b, c -> a, b, c, b, c

Examples

iex> Identicon.reflect_columns([:a, :b, :c])
[:a, :b, :c, :b, :a]

iex> Identicon.reflect_columns(["Jeff", "Dan", 1234])
["Jeff", "Dan", 1234, "Dan", "Jeff"]

runner(input)

Main runner for the project

Examples

iex> Identicon.runner("Andy")
:ok

save_image(image, input)