identicon v0.1.0 Identicon

Creates an Identicon based on an input.

Functions

build_grid(image)

Builds a grid using the hash

build_pixel_map(image)

Builds a pixel map using the grid, returns a tuple of top_left and bottom_right

draw_image(image)

Draws an image using Erlang’s egd. It takes the color and the pixel map

filter_odd_squares(image)

Deletes the odd squares in the grid, the ones which wont be colored

hash_input(input)

Hashes the input using md5

main(input)

Takes input and pipes it through the rest of the helper functions

mirror_row(row)

Helper function, mirrors each row in the grid

pick_color(image)

Picks a color using the first three numbers of the hash

save_image(image, input)

Saves an image using the input given