identicon v0.1.0 Identicon

Identicon application for generating a random identicon based on user input string. Generates an image and saves it to current directory under user_input.png

Docs can be found under doc directory

Functions

build_grid(image)

Builds a 5x5 grid based on hex array of 16 elements. Drops the last element, groups them by 3, mirrors and adds indexes. Sets output to grid property of Identicon.Image struct

build_pixel_map(image)

Builds an image sized 250x250 and coloring the squares based on grid property of the struct Sets output to pixel_map property of Identicon.Image struct

draw_image(image)

Draws the image with given color and pixel_map properties using egd library from Erlang

filter_odd_squares(image)

Filters out odd elements from the given grid, keeping only even numbered ones. Sets output to grid property of Identicon.Image struct

hash_input(input)

Generates a list of lenght 16 consisting of md5 hashed string converted to binaries. Sets the output to hex property of Identicon.Image struct.

Examples

iex> Identicon.hash_input("World")
%Identicon.Image{
  color: nil,
  grid: nil,
  hex: [245, 167, 146, 78, 98, 30, 132, 201, 40, 10, 154, 39, 225, 188, 183,
  246],
  pixel_map: nil
  }

main(input)

Main function of the application

mirror_row(row)

Generates a new center mirrored array from the given three element array by mirroring the first two elements as fourth and fifth

Examples

iex> Identicon.mirror_row([1,2,3])
[1, 2, 3, 2, 1]

pick_color(image)

Picks the first three element of the hex array as color codes for RGB values Sets the output to the color property of Identicon.Image struct

save_image(image, input)

Saves the given image under the user_input.png name