Identicon (identicon v0.1.0)

Provides functions for generating identicons.

Functions

build_image(image)

@spec build_image(Identicon.Image.t()) :: Identicon.Image.t()

Updates the Identicon Image struct by building a new grid from the input hex list.

Examples

iex> Identicon.build_image(%Identicon.Image{hex: [116, 181, 101, 134, 90, 25, 44, 200, 105, 60, 83, 13, 72, 235, 56, 58]})
%Identicon.Image{
  color: nil,
  grid: [{116, 0},
          {181, 1},
          {101, 2},
          {181, 3},
          {116, 4},
          {134, 5},
          {90, 6},
          {25, 7},
          {90, 8},
          {134, 9},
          {44, 10},
          {200, 11},
          {105, 12},
          {200, 13},
          {44, 14},
          {60, 15},
          {83, 16},
          {13, 17},
          {83, 18},
          {60, 19},
          {72, 20},
          {235, 21},
          {56, 22},
          {235, 23},
          {72, 24}],
    hex: [116, 181, 101, 134, 90, 25, 44, 200, 105, 60, 83, 13, 72, 235, 56, 58]}

build_pixel_map(image)

@spec build_pixel_map(Identicon.Image.t()) :: Identicon.Image.t()

Updates the Identicon Image struct with a new pixel_map field.

Examples

iex> Identicon.build_pixel_map(%Identicon.Image{color: [116, 181, 101], grid: [{116, 0}, {134, 3}, {90, 4}, {44, 6}, {200, 7}, {60, 9}, {72, 12}, {56, 14}, {58, 15}]})
%Identicon.Image{
          color: [116, 181, 101],
          grid: [{116, 0}, {134, 3}, {90, 4}, {44, 6}, {200, 7}, {60, 9}, {72, 12}, {56, 14}, {58, 15}],
          hex: nil,
          pixel_map: [
            {{0, 0}, {50, 50}},
            {{150, 0}, {200, 50}},
            {{200, 0}, {250, 50}},
            {{50, 50}, {100, 100}},
            {{100, 50}, {150, 100}},
            {{200, 50}, {250, 100}},
            {{100, 100}, {150, 150}},
            {{200, 100}, {250, 150}},
            {{0, 150}, {50, 200}}
          ]
        }

draw_image(image, input)

@spec draw_image(Identicon.Image.t(), String.t()) :: :ok

Draws the image and saves it to a file with the input string as the file name under the folder ./identicons.{name}.png. Uses the :egd save implements to save the binary file to the location specified.

Examples

iex> Identicon.draw_image(%Identicon.Image{color: [116, 181, 101], pixel_map: [{{0, 0}, {50, 50}}, {{150, 0}, {200, 50}}, {{200, 0}, {250, 50}}, {{50, 50}, {100, 100}}, {{100, 50}, {150, 100}}, {{200, 50}, {250, 100}}, {{100, 100}, {150, 150}}, {{200, 100}, {250, 150}}, {{0, 150}, {50, 200}}]}, "elixir")
:ok

filter_odd_squares(image)

Updates the Identicon Image struct by filtering the grid field to only include the even numbers.

Examples

iex> Identicon.filter_odd_squares(%Identicon.Image{grid: [{116, 0}, {181, 1}, {101, 2}, {134, 3}, {90, 4}, {25, 5}, {44, 6}, {200, 7}, {105, 8}, {60, 9}, {83, 10}, {13, 11}, {72, 12}, {235, 13}, {56, 14}, {58, 15}]})
%Identicon.Image{color: nil, grid: [{116, 0}, {134, 3}, {90, 4}, {44, 6}, {200, 7}, {60, 9}, {72, 12}, {56, 14}, {58, 15}], hex: nil}

hash_input(input)

@spec hash_input(String.t()) :: Identicon.Image.t()

Hashes the input string and returns an updated Identicon.Image struct with the hashed input as the value for the hex field.

Examples

iex> Identicon.hash_input("elixir")
%Identicon.Image{hex: [116, 181, 101, 134, 90, 25, 44, 200, 105, 60, 83, 13, 72, 235, 56, 58]}

main(input)

The main function that takes an input string and returns an Identicon image. Returns an Identicon image, with the :ok atom if the image was successfully saved to a file.

Examples

iex> Identicon.main("elixir")
:ok

mirror_image(row)

@spec mirror_image(list()) :: list()

Mirrors the input list and returns a new list with the mirrored elements appended to the original list.

Examples

iex> Identicon.mirror_image([116, 181, 101])
[116, 181, 101, 181, 116]

pick_color(image)

@spec pick_color(Identicon.Image.t()) :: Identicon.Image.t()

Picks the first three elements from the input list and updates the Identicon Image stuct with the new image field.

## Examples

  iex> Identicon.pick_color(%Identicon.Image{hex: [116, 181, 101, 134, 90, 25, 44, 200, 105, 60, 83, 13, 72, 235, 56, 58]})
  %Identicon.Image{hex: [116, 181, 101, 134, 90, 25, 44, 200, 105, 60, 83, 13, 72, 235, 56, 58], color: [116, 181, 101], grid: nil}