identicon v0.1.0 Identicon

Genera un identicon por un nombre de usuario de parametro

Functions

build_grid(image)

Aplica Enum.chunk a la lista parametro.El resultado de Enum.chunk pasa por Enum.map y a cada row se le aplica la funcion mirror_row. Despues la Lista de listas se junta en una sola con List.flatten. Con Enum.with_index se le agrega el index a cada elemento de la lista.

filter_odd_squares(image)

Toma el grid y solo regresa los elementos pares con rem(code, 2). Se puede poner la sintaxis con una solo linea como se muestra en el ejemplo, o se pueden omitir los primeros parentesis.

Examples

iex> grid = Enum.filter(grid, fn({code, _index}) -> rem(code, 2) == 0 end)

iex> grid = Enum.filter grid, fn({code, _index}) ->
  rem(code, 2) == 0
end

hash_input(input)

Transforma el input string en hash y despues a lista. Para que siempre que se ingrese el mismo string siempre sea el mismo resultado de lista. El resultado es guardado en la estructura Identicon.Image.

main(input)

metodo principal, ejecuta los demas metodos: hash_input, pick_color, build_grid

mirror_row(row)

Metodo para hacer mirror de los primeros dos elementos de la lista. Entrada [1, 2, 3]. Salida [1, 2, 3, 2, 1]. Con ++ se unen listas.

pick_color(image)

Parametro de entrada estructura. Dentro de la estructura hay una lista con 16 elementos pero solo nos interesa los primeros 3 elementos ya que con esos 3 eligiremos el color, se utiliza pattern matching para obtener solo los primeros 3 elementos. El parametro image se puede procesar desde que se recibe (por eso se procesa desde que se recibe en la primera linea de codigo).