identicon v0.1.0 Identicon

The module Identicon is responsible to handle all the functions that transforms a string into a identicon. Identicon is the default profile picture of github. It is a symmetrical image. It's size is 250px/250px containing 25 squares, each square is 50px/50px

Functions

build_grid(image)

This function splits the %Identicon.Image{hex} array into a list of smaller lists cointaining each the length of 3 and then mirrors it. The return of the function will be an array of array, where each small array will have the length of 5

build_pixel_map(image)

This function maps the top_left and bottom_right pixels of each square that will be painted later. The argument is the image struct. The return of this function is image struct with the property pixel_map updated

draw_image(image)

This function creates the Identicon image. The argument of this function is the image struct. The return of this function is the Identicon.

filter_odd_squares(image)

This function filter just the even values of grid and replace it on the struct It recieves the struct image as parameter and return the same struct with the property grid updated. This filter is needed because in our image we will paint just the even values

hash_input(input)

This function transforms the input string in to a hash

main(input)

This function is responsible to call all the other functions in a pipe. It recieves a string as input and returns the identicon

mirror_row(row)

This function recieve the array row as argument and mirrors it's value

pick_color(image)

This function get the first 3 indexes from Identicon.Image{hex} and uses it to create a color in RGB The argument of the function is the struc Identicon.Image and we are immediately Pattern Matchin on the definition of the function The return of the function is the struct with the color property updated.

save_image(image, input)

This function save the Identicon image in the system