Identicon (identicon v0.1.0)

Provides methods for creating a unique identicon based on the user's input.

Functions

build_grid(image)

Builds the image grid and maps it with a reference to the helper function mirror_row

build_pixel_map(image)

Builds the pixel map, over which the Identicon will be rendered

draw_image(image)

Renders the identicon.

filter_odd_squares(image)

Filters out odd squares so they do not get colored.

hash_input(input)

Generates a unique list of numbers based on an input, and calls Indenticon.Image to return it as a struct.

main(input)

Accepts an input and passes it through the sequence of functions to generate the identicon.

mirror_row(row)

Accepts the RGB value for each row and expands it into a palindrome

pick_color(image)

Accepts the hex struct and returns the first three indices as an RGB value, and the rest of the struct as a tail.

save_image(image, input)

Saves the Identicon to a png file that is named after the user's input.